requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
python-multipart==0.0.6 